[WIP]big refactor
