+++ src/Exception/MissingTypeException.php
@@ -1,0 +1,6 @@
+<?php
+
+class MissingTypeException extends Exception
+{
+
+}
