+++ src/TypeFactory/Classificationstore.php
@@ -5,22 +5,24 @@
  * @author      Kamil Janik <kjanik@divante.co>
  */
 
-namespace Divante\GraphQlBundle\Type;
+namespace Divante\GraphQlBundle\TypeFactory;
 
-use Divante\GraphQlBundle\Data\Provider;
+use \Divante\GraphQlBundle\TypeFactory\ICustomTypeFactory;
+use \Divante\GraphQlBundle\DataManagement;
 use GraphQL\Type\Definition\ResolveInfo;
 use Pimcore\Model\DataObject\ClassDefinition\Data;
 use GraphQL\Type\Definition\Type;
 use GraphQL\Type\Definition\ObjectType;
 
 /**
- * Class FieldType
- * @package Divante\GraphQlBundle\Type
+ * Class Classificationstore
  */
-class FieldType
+class Classificationstore implements ICustomTypeFactory
 {
+    const SUPPORTED_TYPE = "\\Pimcore\\Model\\DataObject\\Classificationstore";
+
     /**
-     * @var Provider
+     * @var DataManagement\Query\Classificationstore
      */
     private $dataProvider;
 
@@ -30,113 +32,23 @@ class FieldType
     private $typeList = [];
 
     /**
-     * @param Provider $dataProvider
+     * @param DataManagement\Query\Basic $dataProvider
      * @required
      */
-    public function setDataProvider(Provider $dataProvider)
+    public function setDataProvider(DataManagement\Query\Classificationstore $dataProvider)
     {
         $this->dataProvider = $dataProvider;
     }
 
-    /**
-     * @param Data $fieldDefinition
-     * @return bool
-     */
-    public function isReferenceType(Data $fieldDefinition)
+    public function supports(string $type)
     {
-        return in_array(
-            $fieldDefinition->getFieldtype(),
-            ["objects","manyToManyObjectRelation", "href", "manyToOneRelation"]
-        );
-    }
-
-    /**
-     * @param Data $fieldDefinition
-     * @return bool
-     */
-    public function isUnionType(Data $fieldDefinition)
-    {
-        if (!$this->isReferenceType($fieldDefinition)) {
-            return false;
-        }
-        $classes = $fieldDefinition->getClasses();
-        return count($classes) > 1;
-    }
-
-    /**
-     * @param Data $fieldDefinition
-     * @return bool
-     */
-    public function isCollectionReferenceType(Data $fieldDefinition)
-    {
-        return in_array($fieldDefinition->getFieldtype(), ["objects", "manyToManyObjectRelation"]);
-    }
-
-    /**
-     * @param Data $fieldDefinition
-     * @return mixed
-     */
-    public function getClassName(Data $fieldDefinition)
-    {
-        $classes = $fieldDefinition->getClasses();
-        return $classes[0]["classes"] ?? null;
-    }
-
-    /**
-     * @param Data $fieldDefinition
-     * @return array
-     */
-    public function getClassNameCollection(Data $fieldDefinition)
-    {
-        return [];
-    }
-
-    /**
-     * @param Data $fieldDefinition
-     * @return \GraphQL\Type\Definition\BooleanType|\GraphQL\Type\Definition\FloatType|\GraphQL\Type\Definition\IntType|\GraphQL\Type\Definition\ListOfType|\GraphQL\Type\Definition\StringType
-     */
-    public function getSimpleType(Data $fieldDefinition)
-    {
-        switch ($fieldDefinition->getPhpdocType()) {
-            case "string":
-                return Type::string();
-                break;
-            case "boolean":
-                return Type::boolean();
-                break;
-            case "array":
-                return Type::listOf(Type::string());
-                break;
-            case "float":
-                return Type::float();
-                break;
-            case "int":
-                return Type::int();
-                break;
-            default:
-                return $this->getCustomType($fieldDefinition);
-        }
-    }
-
-    /**
-     * @param Data $fieldDefinition
-     * @return \GraphQL\Type\Definition\IntType|mixed
-     */
-    private function getCustomType(Data $fieldDefinition)
-    {
-        switch ($fieldDefinition->getPhpdocType()) {
-            case "\\Pimcore\\Model\\DataObject\\Classificationstore":
-                return $this->getClassificationstoreType();
-                break;
-            default:
-                return Type::int();
-        }
+        return self::SUPPORTED_TYPE == $type;
     }
 
     /**
      * @return mixed
      */
-    private function getClassificationstoreType()
+    public function getCustomType()
     {
         if (!(($this->typeList['classificationstore'] ?? null) instanceof  ObjectType)) {
             $this->typeList['classificationstore'] = new ObjectType([
