+++ src/TypeFactory/ICustomTypeFactory.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Divante\GraphQlBundle\TypeFactory;
+
+interface ICustomTypeFactory
+{
+    public function supports(string $type);
+    public function getCustomType();
+}
